fsp: Move IMultiCommitManager to a seperate file
